Create Resusable components from existing code and refactor folder structure
