+++ lib/views/screens/auth/forgot_password_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:velo_toulouse_redesign/view_model/user_view_model.dart';
+import 'package:velo_toulouse_redesign/view_model/user_viewmodel.dart';
 
 class ForgotPasswordScreen extends ConsumerStatefulWidget {
   const ForgotPasswordScreen({super.key});
