+++ lib/views/screens/bike_renting_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:velo_toulouse_redesign/core/theme/theme.dart';
 import 'package:velo_toulouse_redesign/data/models/bike_model.dart';
 import 'package:velo_toulouse_redesign/views/screens/payment_screen.dart';
+import 'package:velo_toulouse_redesign/views/widgets/bottom_action_container.dart';
 import 'package:velo_toulouse_redesign/views/widgets/buttons/button.dart';
 import 'package:velo_toulouse_redesign/views/widgets/top_bar/app_bar.dart';
 
@@ -324,23 +325,15 @@ class _BikeRentingScreenState extends State<BikeRentingScreen> {
           ),
 
           // ── Bottom CTA ────────────────
-          Container(
-            padding: const EdgeInsets.fromLTRB(20, 12, 20, 28),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withValues(alpha: 0.06),
-                  blurRadius: 16,
-                  offset: const Offset(0, -4),
-                ),
-              ],
-            ),
+          BottomActionContainer(
             child: VeloButton(
               text: "Pay now",
               onPressed: () => Navigator.push(
                 context,
-                MaterialPageRoute(builder: (context) => PaymentScreen(plateNumber: widget.bike.plateNumber,)),
+                MaterialPageRoute(
+                  builder: (context) =>
+                      PaymentScreen(plateNumber: widget.bike.plateNumber),
+                ),
               ),
             ),
           ),
