+++ lib/views/screens/ride_summary_screen.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:velo_toulouse_redesign/core/theme/theme.dart';
 import 'package:velo_toulouse_redesign/data/models/station_model.dart';
+import 'package:velo_toulouse_redesign/views/widgets/bottom_action_container.dart';
 import 'package:velo_toulouse_redesign/views/widgets/buttons/button.dart';
+import 'package:velo_toulouse_redesign/views/widgets/success_header.dart';
 
 class RideSummaryScreen extends StatelessWidget {
   final String bikeNumber;
@@ -33,37 +35,13 @@ class RideSummaryScreen extends StatelessWidget {
               padding: const EdgeInsets.fromLTRB(20, 60, 20, 20),
               child: Column(
                 children: [
-                  // ── Success icon ──────────────────────────────────
-                  Container(
-                    width: 80,
-                    height: 80,
-                    decoration: const BoxDecoration(
-                      color: Color(0xFF006D33),
-                      shape: BoxShape.circle,
-                    ),
-                    child: const Icon(
-                      Icons.check_rounded,
-                      color: Colors.white,
-                      size: 44,
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-
-                  Text(
-                    'Bike Returned!',
-                    style: AppTextStyles.heading.copyWith(
-                      fontSize: 26,
-                      fontWeight: FontWeight.w700,
-                      color: const Color(0xFF1A1A1A),
-                    ),
-                  ),
-                  const SizedBox(height: 6),
-                  Text(
-                    'Thanks for riding with Vélo Toulouse',
-                    style: AppTextStyles.label.copyWith(
-                      color: Colors.grey[500],
-                      fontSize: 14,
-                    ),
+                  const SuccessHeader(
+                    title: 'Bike Returned!',
+                    subtitle: 'Thanks for riding with Vélo Toulouse',
+                    circleColor: Color(0xFF006D33),
+                    iconColor: Colors.white,
+                    circleSize: 80,
+                    iconSize: 44,
                   ),
 
                   const SizedBox(height: 36),
@@ -146,18 +124,7 @@ class RideSummaryScreen extends StatelessWidget {
           ),
 
           // ── Bottom CTA ────────────────────────────────────────────
-          Container(
-            padding: const EdgeInsets.fromLTRB(20, 12, 20, 28),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withValues(alpha: 0.06),
-                  blurRadius: 16,
-                  offset: const Offset(0, -4),
-                ),
-              ],
-            ),
+          BottomActionContainer(
             child: VeloButton(
               text: 'Back to Map',
               onPressed: () {
