+++ lib/views/widgets/legend_pill.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'package:velo_toulouse_redesign/core/theme/theme.dart';
+
+class LegendPill extends StatelessWidget {
+  final String text;
+
+  const LegendPill({super.key, required this.text});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 8),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withValues(alpha: 0.08),
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Container(
+            width: 10,
+            height: 10,
+            decoration: const BoxDecoration(
+              shape: BoxShape.circle,
+              color: AppColors.primaryColor,
+            ),
+          ),
+          const SizedBox(width: 6),
+          Text(
+            text,
+            style: AppTextStyles.label.copyWith(
+              fontSize: 12,
+              color: Colors.grey[600],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
