+++ lib/views/widgets/ride_bottom_sheet.dart
@@ -1,0 +1,183 @@
+import 'package:flutter/material.dart';
+import 'package:velo_toulouse_redesign/core/theme/theme.dart';
+import 'package:velo_toulouse_redesign/data/models/station_model.dart';
+import 'package:velo_toulouse_redesign/views/widgets/swipe_to_dock.dart';
+
+class RideBottomSheet extends StatelessWidget {
+  final String formattedTime;
+  final String bikeNumber;
+  final StationModel? returnStation;
+  final bool hasReturnStation;
+  final VoidCallback onDocked;
+  final VoidCallback onClearReturn;
+
+  const RideBottomSheet({
+    super.key,
+    required this.formattedTime,
+    required this.bikeNumber,
+    required this.returnStation,
+    required this.hasReturnStation,
+    required this.onDocked,
+    required this.onClearReturn,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: const BoxDecoration(
+        color: Color(0xFF006D33),
+        borderRadius: BorderRadius.vertical(top: Radius.circular(28)),
+      ),
+      padding: EdgeInsets.fromLTRB(
+        20,
+        16,
+        20,
+        MediaQuery.of(context).padding.bottom + 20,
+      ),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Container(
+            width: 36,
+            height: 4,
+            decoration: BoxDecoration(
+              color: Colors.white.withValues(alpha: 0.3),
+              borderRadius: BorderRadius.circular(2),
+            ),
+          ),
+          const SizedBox(height: 20),
+          Row(
+            children: [
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Elapsed time',
+                      style: AppTextStyles.label.copyWith(
+                        color: Colors.white54,
+                        fontSize: 12,
+                      ),
+                    ),
+                    const SizedBox(height: 2),
+                    Text(
+                      formattedTime,
+                      style: AppTextStyles.heading.copyWith(
+                        color: Colors.white,
+                        fontSize: 38,
+                        fontWeight: FontWeight.w700,
+                        letterSpacing: 1.5,
+                        height: 1,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              Container(
+                width: 1,
+                height: 48,
+                color: Colors.white.withValues(alpha: 0.2),
+              ),
+              const SizedBox(width: 20),
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    'Bike in use',
+                    style: AppTextStyles.label.copyWith(
+                      color: Colors.white54,
+                      fontSize: 12,
+                    ),
+                  ),
+                  const SizedBox(height: 2),
+                  Row(
+                    children: [
+                      const Icon(
+                        Icons.directions_bike_rounded,
+                        color: Colors.white,
+                        size: 18,
+                      ),
+                      const SizedBox(width: 6),
+                      Text(
+                        bikeNumber,
+                        style: AppTextStyles.body.copyWith(
+                          color: Colors.white,
+                          fontSize: 16,
+                          fontWeight: FontWeight.w700,
+                          letterSpacing: 1,
+                        ),
+                      ),
+                    ],
+                  ),
+                ],
+              ),
+            ],
+          ),
+          if (hasReturnStation) ...[
+            const SizedBox(height: 16),
+            Container(
+              padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 10),
+              decoration: BoxDecoration(
+                color: Colors.white.withValues(alpha: 0.12),
+                borderRadius: BorderRadius.circular(12),
+              ),
+              child: Row(
+                children: [
+                  const Icon(Icons.flag_rounded, color: Colors.white, size: 16),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: Text(
+                      returnStation!.name,
+                      style: AppTextStyles.label.copyWith(
+                        color: Colors.white,
+                        fontSize: 13,
+                        fontWeight: FontWeight.w600,
+                      ),
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                  ),
+                  GestureDetector(
+                    onTap: onClearReturn,
+                    child: Text(
+                      'Change',
+                      style: AppTextStyles.label.copyWith(
+                        color: Colors.white60,
+                        fontSize: 12,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+          const SizedBox(height: 16),
+          hasReturnStation
+              ? SwipeToDock(onDocked: onDocked)
+              : SizedBox(
+                  width: double.infinity,
+                  child: OutlinedButton(
+                    onPressed: null,
+                    style: OutlinedButton.styleFrom(
+                      side: BorderSide(
+                        color: Colors.white.withValues(alpha: 0.4),
+                        width: 1,
+                      ),
+                      padding: const EdgeInsets.symmetric(vertical: 14),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(14),
+                      ),
+                    ),
+                    child: Text(
+                      'Tap a station on the map to return',
+                      style: AppTextStyles.label.copyWith(
+                        color: Colors.white60,
+                        fontSize: 14,
+                      ),
+                    ),
+                  ),
+                ),
+        ],
+      ),
+    );
+  }
+}
